+++ modules/perception/base/image_8u.h
@@ -65,14 +65,14 @@ class Image8U {
     CHECK_EQ(blob_->shape(2), channels_);
     CHECK_LE(offset_ + blob_->offset({rows - 1, cols - 1, channels_ - 1}),
              blob_->count());
-    width_step_ = blob_->offset({1, 0, 0}) * sizeof(uint8_t);
+    width_step_ = blob_->offset({1, 0, 0}) * static_cast<int>(sizeof(uint8_t));
   }
 
   Image8U(int rows, int cols, Color type)
       : rows_(rows), cols_(cols), type_(type), offset_(0) {
     channels_ = kChannelsMap.at(type);
     blob_.reset(new Blob<uint8_t>({rows_, cols_, channels_}));
-    width_step_ = blob_->offset({1, 0, 0}) * sizeof(uint8_t);
+    width_step_ = blob_->offset({1, 0, 0}) * static_cast<int>(sizeof(uint8_t));
   }
 
   Image8U(const Image8U &src)
